requirements.txt
streamlit>=1.32.0
google-generativeai>=0.5.0
sentence-transformers>=2.6.0
faiss-cpu>=1.8.0
python-dotenv>=1.0.0
numpy>=1.26.0
